Moved custom animation hooks to a separate file
